CIT-56 vente Managments

+++ src/main/java/com/citronix/citronix/repository/VenteRepository.java
@@ -3,5 +3,5 @@ package com.citronix.citronix.repository;
 import com.citronix.citronix.entity.Vente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VenteRepository extends JpaRepository<Vente, Integer> {
+public interface VenteRepository extends JpaRepository<Vente, Long> {
 }
